third_party/blink/web_tests/fast/canvas-api/canvas-createImageBitmap-invalid-args-in-workers-expected.txt

[Worker] Test createImageBitmap with invalid arguments in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ./resources/canvas-createImageBitmap-invalid-args-in-workers.js
PASS [Worker] Rejected as expected: null
PASS [Worker] reason instanceof Error is true
[Worker] TypeError: Failed to execute 'createImageBitmap' on 'WorkerGlobalScope': The provided value is not of type '(Blob or HTMLCanvasElement or HTMLImageElement or HTMLVideoElement or ImageBitmap or ImageData or OffscreenCanvas or SVGImageElement or VideoFrame)'.
PASS [Worker] Rejected as expected: invalid area
PASS [Worker] reason instanceof Error is true
[Worker] RangeError: Failed to execute 'createImageBitmap' on 'WorkerGlobalScope': The crop rect height is 0.
PASS [Worker] Rejected as expected: invalid area
PASS [Worker] reason instanceof Error is true
[Worker] RangeError: Failed to execute 'createImageBitmap' on 'WorkerGlobalScope': The crop rect width is 0.
PASS successfullyParsed is true

TEST COMPLETE
